requirements.txt
chainlit>=1.0.0
langchain>=0.2.0
ollama>=0.1.0
passlib
bcrypt
python-dotenv
uvicorn